updated to create tables if they don't exist

+++ src/main/java/ua/lsi/media_tracker/model/Media.java
@@ -11,16 +11,13 @@ import java.text.NumberFormat;
  * @author LSI
  */
 @Data
-//@Entity
 @Builder
 @EqualsAndHashCode(exclude = "section")
 @NoArgsConstructor
 @AllArgsConstructor
 public class Media {
 
-//    @Id
-//    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
+    private Integer id;
 
     private final static NumberFormat NUMBER_FORMAT = NumberFormat.getInstance();
 
@@ -32,8 +29,6 @@ public class Media {
 
     private int position;
 
-//    @ManyToOne
-//    @JoinColumn(name = "section_id")
     private Section section;
 
     public void change(String property, String sign) {
